Refactor TextField component styles and stories
This commit refactors the styles and stories of the TextField component. It reorders the CSS classes in the TextField module and updates the TextField.stories.tsx file to include a new prop, 'isCode'. Additionally, it adds a new description prop to the Invalid story in the TextField.stories.tsx file.